+++ R/add.R
@@ -218,7 +218,9 @@ addAssays <- function(study, assays, reset = FALSE) {
 #'   modelIDs. For each modelID, each element of the nested list is a test. The
 #'   names should be the testIDs. The value should be a single character string
 #'   describing the testID. To share tests across multiple models, use the
-#'   modelID "default".
+#'   modelID "default". Instead of a single character string, you can provide a
+#'   list of metadata fields about each test. The field "description" will be
+#'   used to derive the tooltip displayed in the app.
 #' @inheritParams shared-add
 #'
 #' @examples